+++ src/pytest_logging_strict/plugin.py
@@ -107,6 +107,10 @@ class LSConfigStash:
 
         return cls(logging_strict_config_yaml_path=path_f)
 
+    def remove(self):
+        """unlink file"""
+        self.logging_strict_config_yaml_path.unlink(missing_ok=True)
+
     def serialized(self):
         """Getter
 
@@ -192,24 +196,77 @@ def pytest_configure(config):
 
     Configure the plugin based on the CLI.
 
+    :py:mod:`tempfile` does not automatically flush. Contents in a buffer
+    upon close writes buffer to file.
+
+    pypy correctly doesn't see contents in file cuz its in buffer until
+    the file is closed. Rather than wait for a close, flush the buffer
+    to file.
+
+    >>> from pathlib import Path
+    >>> from tempfile import NamedTemporaryFile
+    >>> f = NamedTemporaryFile(delete=False)
+    >>> f.write(b"Hello World!")
+    12
+    >>> Path(f.name).stat().st_size
+    0
+    >>> f.flush()
+    >>> Path(f.name).stat().st_size
+    12
+
     :param config: pytest configuration
     :type config: pytest.Config
+
+    .. todo:: Change base folder
+
+       Runner is pytest, not :py:class:`tempfile.NamedTemporaryFile`.
+       NamedTemporaryFile base folder is ``/tmp``. Should the base
+       folder be in a pytest determined location?
+
     """
     xdist_worker = _xdist_worker(config)
     key = stash_key["config"]
     if not xdist_worker:
         config.pluginmanager.register(LoggingStrictControllerPlugin())
 
+        """
+        import os
+        env_cov = os.environ.get("COVERAGE_PROCESS_START", None)
+        if env_cov is not None and isinstance(env_cov, str) and len(env_cov) != 0:
+            import coverage
+
+            coverage.process_startup()
+        else:  # pragma: no cover
+            pass
+        """
+        pass
+
+        """tempfile.NamedTemporaryFile temp folder is ``/tmp``.
+        Instead use _pytest.tmpdir.TempPathFactory.getbasetemp which is
+        similiar to session scoped
+
+        given_basetemp = None
+        retention_count = 0
+        temppath_factory = pytest.TempPathFactory(
+            given_basetemp,
+            retention_count,
+            retention_policy="all",
+            trace=config.trace.get("tmpdir"),
+        )
+        path_dir = temppath_factory.mktemp("session-", numbered=True)
         # Get the path to a temporary file
+        f = NamedTemporaryFile(delete=False, dir=path_dir)
+        """
         f = NamedTemporaryFile(delete=False)
         path_f = Path(f.name)
 
         """Search for logging config YAML file. Validate then write
-        contents to temp file"""
+        contents to buffer then flush to temp file"""
         str_yaml = get_yaml(config, path_f)
         if str_yaml is not None:
             bytes_yaml = str_yaml.encode("utf-8")
             f.write(bytes_yaml)
+            f.flush()
         else:
             # Write warning
             msg_warn = (
@@ -245,14 +302,13 @@ class LoggingStrictControllerPlugin:
     """pytest plugin for logging-strict"""
 
     def pytest_unconfigure(self, config):
-        """Clean up the mypy results path.
+        """Clean up temp file tracked by pytest stash.
 
         :param config: pytest configuration
         :type config: pytest.Config
         """
-        config.stash[stash_key["config"]].logging_strict_config_yaml_path.unlink(
-            missing_ok=True,
-        )
+        stash_inst = config.stash[stash_key["config"]]
+        stash_inst.remove()
 
 
 def pytest_addoption(parser):
@@ -292,22 +348,23 @@ def logging_strict_get(logging_strict_get_stash_path: pytest.Fixture):
     logging config YAML str.
 
     :returns: logging config YAML str
-    :rtype: str | None
+    :rtype: tuple[str | None, pathlib.Path]
     """
     path_ls_yaml_f = logging_strict_get_stash_path
 
     try:
         bytes_yaml_raw = path_ls_yaml_f.read_bytes()
         str_yaml_raw = bytes_yaml_raw.decode()
-    except OSError as exc:
+    except (OSError, Exception) as exc:
+        # catch-all Exception not expected
         msg_warn = (
-            "Could not retrieve logging config YAML from "
+            "Could not read extracted logging config YAML from "
             f"temp file {path_ls_yaml_f!r} {exc!r}"
         )
         warnings.warn(msg_warn)
         str_yaml_raw = None
 
-    return str_yaml_raw
+    return (str_yaml_raw, path_ls_yaml_f)
 
 
 @pytest.fixture
@@ -346,7 +403,8 @@ def get_d_config(
             logging_package_name = None
 
         # Get path, to logging config YAML file, from pytest stash
-        str_yaml_raw = logging_strict_get
+        t_two = logging_strict_get
+        str_yaml_raw, path_f = t_two
         is_search_fail = (
             str_yaml_raw is None
             or not isinstance(str_yaml_raw, str)
@@ -355,7 +413,9 @@ def get_d_config(
         if is_search_fail:  # pragma: no cover
             msg_warn = (
                 "Search for a matching logging config YAML file failed. "
-                "Either package contains none or need to widen the search"
+                f"Extracted temp file size: {path_f.stat().st_size} {path_f!r}"
+                "Either package lacks these files or need to widen the "
+                "search (parameters)"
             )
             warnings.warn(msg_warn)
             ret = None